requirements.txt
cryptography
python-dateutil
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug
PyMySQL
python-dotenv
yfinance
pycoingecko
Flask-WTF
scikit-learn
pandas
numpy
gunicorn
psycopg2-binary
Flask-Migrate
joblib
requests
prophet
fpdf2
torch
torchvision
torchaudio
easyocr
transformers
accelerate
Pillow
opencv-python
regex
sentencepiece
google-generativeai
cloudinary